+++ cli/mf_extract_wrapper.js
@@ -4,8 +4,6 @@ var path	= require('path');
 var spawn = require('child_process').spawn;
 var projRoot, cli, args;
 
-var cliPath = 'packages/messageformat/cli/mf_extract.js';
-
 for (projRoot = process.cwd();
 	projRoot != '/' && !fs.existsSync(projRoot + '/.meteor/release');
 	projRoot = path.normalize(projRoot + '/..'));
@@ -15,14 +13,52 @@ if (!fs.existsSync(projRoot + '/.meteor/release')) {
 	process.exit(1);
 }
 
-if (!fs.existsSync(projRoot + '/' + cliPath)) {
-	console.log("Error: Can't find " + cliPath + "\n"
-		+ "Are you sure messageformat is installed in this project?");
-	process.exit(1);
+var possibleNames = ['messageformat', 'gadicohen:messageformat',
+	'mf:core', 'messageformat:core'];  // possible future names
+
+var cliPath = null;
+var npmBuild = null;
+
+for (var name, i=0; i < possibleNames.length; i++) {
+	name = possibleNames[i];
+	if (fs.existsSync(path.join(projRoot, 'packages', name))) {
+		cliPath = path.join(projRoot, 'packages', name, 'cli', 'mf_extract.js');
+		npmBuild = path.join(projRoot, 'packages', name, '.build', 'npm', 'node_modules');
+		break;
+	}
 }
 
+if (!cliPath) {
+	var versions = loadVersions(projRoot);
+
+	if (!versions) {
+		// In Meteor < 0.9, if it's not in package directory, it's not installed
+		console.log("Error: Can't find " + cliPath + "\n"
+			+ "Are you sure messageformat is installed in this project?");
+		process.exit(1);
+	}
+
+	for (var name, i=0; i < possibleNames.length; i++) {
+		name = possibleNames[i];
+		if (versions[name]) {
+			cliPath = path.join(process.env.HOME, '.meteor', 'packages', name,
+				versions[name], 'os', 'packages', name, 'cli', 'mf_extract.js');
+			npmBuild = path.join(process.env.HOME, '.meteor', 'packages', name,
+				versions[name], 'npm', 'node_modules');
+			break;
+		}
+	}
+
+	if (!cliPath) {
+		console.log("Error: Messageformat not installed in this project.");
+		process.exit(1);		
+	}
+}
+
+console.log('Using ' + cliPath);
+
 args = process.argv;
-args.splice(0, 2, cliPath);
+args.splice(0, 2, cliPath, npmBuild);
 
 cli = spawn('node', args, {
 	cwd: projRoot,
@@ -33,3 +69,25 @@ cli.stderr.on('data', function(data) { process.stderr.write(data); });
 cli.on('exit', function(code) {
 	process.exit(code);
 });
+
+/* Meteor 0.9 support functions */
+
+function loadVersions(appDir) {
+  var versions = {};
+  var versionsFile = path.join(appDir, '.meteor', 'versions');
+  if (fs.existsSync(versionsFile)) {
+    var raw = fs.readFileSync(versionsFile, 'utf8');
+    var lines = raw.split(/\r*\n\r*/);
+    for (var i=0; i < lines.length; i++) {
+      var line = lines[i].trim();
+      if (/^#/.test(line)) {
+        // Noop we got a comment
+      } else if (line.length) {
+        line = line.split('@');
+        versions[line[0]] = line[1];
+      }
+    }
+	  return versions;
+  }
+  return null;
+}